Split up GTFS.Rmd into multiple different files, now each called from Execute.R.
getBusArrivalData.R basically grabs hourly data for bus arrival times based on when it was called. Set it up to create subfolders for time under folder for day, depending on when the file was ran (example in rawdata).

getBusInfoData.R basically runs the reading in on the status bus info txt files (but could be updated to read straight from AT developer portal?)

dataCleansing.Rmd and dataExploration.Rmd are self-explanatory.